last update

+++ models/User.go
@@ -4,10 +4,10 @@ import "gorm.io/gorm"
 
 type User struct {
 	gorm.Model
-	Name     string `json:"name"`
-	Surname  string `json:"surname"`
-	Email    string `gorm:"unique" json:"email"`
-	Password string `json:"password"`
-	Role     string `json:"role"`
-	Balance  int    `json:"balance"`
+	Name     string  `json:"name"`
+	Surname  string  `json:"surname"`
+	Email    string  `gorm:"unique" json:"email"`
+	Password string  `json:"password"`
+	Role     string  `json:"role"`
+	Balance  float64 `json:"balance"`
 }
